+++ src/SauceDemoAutomation/Pages/CartPage.cs
@@ -2,7 +2,7 @@ using OpenQA.Selenium;
 using Serilog;
 using System.Collections.Generic;
 
-namespace TakealotAutomation.Pages
+namespace SauceDemoAutomation.Pages
 {
     /// <summary>
     /// Page object for Sauce Demo shopping cart