+++ src/SauceDemoAutomation/Tests/CheckoutTests.cs
@@ -1,8 +1,8 @@
 using NUnit.Framework;
 using Serilog;
-using TakealotAutomation.Pages;
+using SauceDemoAutomation.Pages;
 
-namespace TakealotAutomation.Tests
+namespace SauceDemoAutomation.Tests
 {
     /// <summary>
     /// Test suite for Sauce Demo checkout and order completion